Switched from MM:SS to HH:MM
Having 3 hands is one too many, especially on a smaller screen. Having
minutes and seconds is also a bit inconvenient, because the minute hand
moves too slowly and makes relative comparisons difficult. So, I've
chosen instead to have hours and minutes, but I'll make minutes tick as
fast as seconds. At least in my mind it's the least confusing option.