Cancel Doppler visualization
The current view is "as is", i.e. a simultenaity plane of the ship. The
Doppler effect, on the other hand, relates to how the stars are seen
when the light from them reaches the ship. Given the positions of the
stars and their clocks will be different in an "as seen" view, I've
moved the idea for Doppler visualization into a possible "as seen" view
extension.